Decoupled Contact Page, State & Model. Updated the notes in README.md
