+++ alembic/versions/7e1bde767718_.py
@@ -1,0 +1,38 @@
+"""empty message
+
+Revision ID: 7e1bde767718
+Revises: 3756679b9a79
+Create Date: 2025-02-24 22:09:33.636423
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+import sqlmodel
+
+# revision identifiers, used by Alembic.
+revision: str = '7e1bde767718'
+down_revision: Union[str, None] = '3756679b9a79'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('contactentrymodel')
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('contactentrymodel',
+    sa.Column('id', sa.INTEGER(), nullable=False),
+    sa.Column('first_name', sa.VARCHAR(), nullable=False),
+    sa.Column('last_name', sa.VARCHAR(), nullable=True),
+    sa.Column('email', sa.VARCHAR(), nullable=True),
+    sa.Column('message', sa.VARCHAR(), nullable=False),
+    sa.Column('created_at', sa.DATETIME(), server_default=sa.text('(CURRENT_TIMESTAMP)'), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    # ### end Alembic commands ###
