+++ full_stack_python/full_stack_python.py
@@ -2,10 +2,11 @@
 
 import reflex as rx
 
+
 from rxconfig import config
 from .ui.base import base_page
-# from .pages.about import about_page
-from . import navigation, pages
+
+from . import navigation, pages, contact
 
 
 class State(rx.State):
@@ -46,3 +47,4 @@ def index() -> rx.Component:
 app = rx.App()
 app.add_page(index)
 app.add_page(pages.about_page, route=navigation.routes.ABOUT_ROUTE)
+app.add_page(contact.contact_page, route=navigation.routes.CONTACT_ROUTE)
