+++ full_stack_python/pages/__init__.py
@@ -1,8 +1,6 @@
 from .about import about_page
-from .contact import contact_page
 
 
 __all__ = [
     'about_page',
-    'contact_page',
 ]
